requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1

# Multi-tenancy
django-tenants==3.5.0

# Database
psycopg2-binary==2.9.9

# Task Queue
celery[redis]==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# UI & Forms
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-htmx==1.17.2

# File handling & PDF
Pillow==10.2.0
reportlab==4.0.9
WeasyPrint==60.2

# Payment gateways
stripe==7.9.0
razorpay==1.4.1

# Email & SMS
django-anymail==10.2
twilio==8.11.1

# Security & Auth
python-decouple==3.8
django-allauth==0.57.0

# Development & Testing
django-debug-toolbar==4.2.0
factory-boy==3.3.0
faker==22.1.0

# API Documentation
drf-spectacular==0.27.0

# Monitoring & Logging
sentry-sdk==1.39.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3